Restructured root and App; Moved loading view to separate component
